feat: refactor calendar command by moving utility functions to a new module
